Enhances intro view and splash screen transitions
Refactors the intro screen to use a `Container` with a background image for improved styling flexibility. Updates the splash screen to include a custom slide-and-fade transition to the intro screen for smoother navigation.

Adds new dependencies (`country_code_picker`, `phone_numbers_parser`) and updates asset paths in `pubspec.yaml` for better resource management.